+++ CHANGELOG.md
@@ -1,4 +1,24 @@
 # Changelog
+## Version 0.6.0 - 2014-10-13
+
+### Added
+- Assert method for Inheritance JPA annotation on class
+- Assert method for Table JPA annotation on class
+- Assert method for MapKeyColumn JPA annotation on field
+- Assert method for AttributeOverrides JPA annotation on field
+- Assert method for AttributeOverride JPA annotation contained inside AttributeOverrides JPA annotation
+- Custom class containing the expected fields of the AttributeOverride JPA annotation
+
+### Deprecated
+- Nothing
+
+### Removed
+- Nothing
+
+### Fixed
+- Assert method for CollectionTable JPA annotation: Made uniqueConstraints length to be less or equal with 1
+
+
 
 ## Version 0.5.0 - 2014-10-13